A multitude amount of progress was made. Titles were changed for a few sections for clarity. Also added a title for the 20 eigenfaces visualisation for the same reason.
In calculating the coordinate vectors section, I have forgotten to include an initialisation for c_vectors in the for loop and so it was included. In the same section of the for loop, the 1000 number was replaced by an existing variable of the same value for simplification. Furthermore, the visualisations in the same section were replaced with a temporary test sub section to verify if the eigenfaces in product with a random coordinate vector did infact produce a good reconstruction of the original random face; which it did. 

For the moustache detector, the initial method was entirely changed with a more efficient plan of displaying moutsache detected faces into one tiled plot. Though the method of filtering coord vectors using a moustache level is the same. The method was accomplished using a mask to generate a matrix of coord vectors that satisfy the moustache level. The corresponding coord vectors of the photo columns in A were then extracted and visualised on a tiled plot. The if statement was developed to experiment with different values of moustache levels without errors occuring in producing the moustache detected face plot. Finally, a temporary test section was made to visualise a specific number of faces, visualise unique faces and calculating and plotting unique faces that satisfy the moustache level. These test sections were made to better understand the face images given.

Simplified mean face calculation by removing a '.' as it is not needed.